requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
geopandas>=0.14.0
leafmap>=0.28.0
folium>=0.15.0
reportlab>=4.0.8
Pillow>=10.0.0
selenium>=4.16.0
webdriver-manager>=4.0.1
streamlit-lottie>=0.0.5
docx2pdf>=0.1.8
docx-mailmerge>=0.5.0
pyodbc>=5.0.1
python-docx>=1.0.1
openpyxl>=3.1.2
xlrd>=2.0.1
python-dotenv>=1.0.0
pathlib>=1.0.1
datetime>=5.3
uuid>=1.30
pythoncom>=2.7
PyYAML>=6.0.1
streamlit
snowflake-connector-python
pandas
pyodbc
sqlalchemy
